Rename bun-index to index
